+++ CHANGELOG.md
@@ -15,12 +15,14 @@ Changelog
 
 **Next release**
 
+[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
+
+**Version 2.0.38 STABLE**
+
 * Showing enabled (on/off) state
 * Requesting restart for *load* and *loadLibrary*
 * Resetting on demand restricting for system applications on upgrade
 * Updated Slovak translation
-
-[Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.0.37 TEST**